Reorganize code into various subdirectories
There is no changes to code, just re-organizaation
