+++ pyproject.toml
@@ -24,7 +24,7 @@ pydantic = "^1.8.1"
 lint = ["flake8", "black", "mypy"]
 
 [tool.poetry.scripts]
-vaccine-feed-ingest = "vaccine_feed_ingest.run:cli"
+vaccine-feed-ingest = "vaccine_feed_ingest.cli:cli"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
