+++ vaccine_feed_ingest/stages/common.py
@@ -1,0 +1,32 @@
+"""Shared constants and methods for running ingest"""
+
+import enum
+import pathlib
+
+# Base directory that stores the code for each site runner
+RUNNERS_DIR = pathlib.Path(__file__).parent.parent / "runners"
+
+
+@enum.unique
+class PipelineStage(str, enum.Enum):
+    """Stages of ingestion pipeline."""
+
+    FETCH = "fetch"
+    PARSE = "parse"
+    NORMALIZE = "normalize"
+
+
+# Root name for command or config to run for each stage e.g. fetch.py
+STAGE_CMD_NAME = {
+    PipelineStage.FETCH: "fetch",
+    PipelineStage.PARSE: "parse",
+    PipelineStage.NORMALIZE: "normalize",
+}
+
+
+# Directory name for where to store data for each stage
+STAGE_OUTPUT_NAME = {
+    PipelineStage.FETCH: "raw",
+    PipelineStage.PARSE: "parsed",
+    PipelineStage.NORMALIZE: "normalized",
+}
